+++ src/board/Game.java
@@ -1,16 +1,34 @@
 package board;
 
-public class Game {
+import javax.swing.*;
+import javax.swing.border.EmptyBorder;
+import java.awt.*;
 
-    public static void main (String[]args){
-        int rows = 4;
-        int cols = 4;
-        Board board = new Board(rows, cols);
-        board.createBoard();
-        for(int x = 0; x < rows; x++){
-            for(int y = 0; y < cols; y++){
-                System.out.print(board.cardLocator(x, y).getName()+ ", ");
-            }
-        }
+public class Game extends JFrame {
+
+    private JPanel board = new JPanel(new BorderLayout());
+    private int rows = 4;
+    private int cols = 4;
+    private BoardView boardView = new BoardView(rows, cols);
+    private final JPanel panel = new JPanel(new BorderLayout());
+
+    public Game(){
+        setTitle("MEMORY GAME");
+        setSize(900, 750);
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setLocation(340, 80);
+       // Border border = BorderFactory.createEmptyBorder(20, 20, 20, 20);
+        panel.setBorder(new EmptyBorder(20, 20, 20, 5));
+
+
+
+
+        panel.add(boardView);
+        add(panel, BorderLayout.CENTER);
     }
+
+    public static void main(String[] args) {
+        new Game().setVisible(true);
+    }
+
 }
